Increased the maximum dataset size. A larger dataset would require more code changes. Also increased dataset size far enough to take half a second to run.

+++ src/dataset.rs
@@ -1,4 +1,4 @@
-pub fn spiral(samples: u8, classes: u8) -> (Vec<[f64; 2]>, Vec<u8>) {
+pub fn spiral(samples: u16, classes: u16) -> (Vec<[f64; 2]>, Vec<u16>) {
     let mut data = Vec::with_capacity(usize::from(samples) * usize::from(classes));
     let mut labels = Vec::with_capacity(usize::from(samples) * usize::from(classes));
